split bf16 test to other files.
